Moved builder into internal package.
