+++ parser/parser/internal/builder_test.go
@@ -1,4 +1,4 @@
-package parser
+package internal
 
 import (
 	"testing"
